fix broken tests after properties migration.
- TODO: maybe we should set the properties upon creating of the coded object?? just in case?? nope... let's delete all the system.getproperties instead :D
