+++ src/test/java/testPackage/JSONFileManagerTests.java
@@ -1,15 +1,16 @@
 package testPackage;
 
+import com.shaft.driver.SHAFT;
 import com.shaft.tools.io.JSONFileManager;
 import com.shaft.validation.Validations;
 import org.testng.annotations.Test;
 
-public class Test_JSONFileManager {
+public class JSONFileManagerTests {
     JSONFileManager testDataJSON;
 
     @Test
     public void readStringFromJson() {
-        testDataJSON = new JSONFileManager(System.getProperty("testDataFolderPath") + "simpleJSON.json");
+        testDataJSON = new JSONFileManager(SHAFT.Properties.paths.properties() + "simpleJSON.json");
         Validations.assertThat().object(testDataJSON.getTestData("x.name")).equals("Mohab Mohie");
     }
 }
