+++ src/test/java/testPackage/JsonActionsTests.java
@@ -1,5 +1,6 @@
 package testPackage;
 
+import com.shaft.driver.SHAFT;
 import org.json.JSONException;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
@@ -14,13 +15,13 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 
-public class Test_JsonActions {
+public class JsonActionsTests {
     @Test
     public void f() throws IOException, ParseException, JSONException {
         String expectedResponse = new String(
-                Files.readAllBytes(Paths.get(System.getProperty("testDataFolderPath") + "JsonFileTest2.json")));
+                Files.readAllBytes(Paths.get(SHAFT.Properties.paths.properties() + "JsonFileTest2.json")));
         String actualResponse = new String(
-                Files.readAllBytes(Paths.get(System.getProperty("testDataFolderPath") + "JsonFileTest.json")));
+                Files.readAllBytes(Paths.get(SHAFT.Properties.paths.properties() + "JsonFileTest.json")));
 
         JSONObject expectedJsonObject = (JSONObject) (new JSONParser()).parse(expectedResponse);
         JSONObject actualJsonObject = (JSONObject) (new JSONParser()).parse(actualResponse);
@@ -37,9 +38,9 @@ public class Test_JsonActions {
     @Test
     public void f2() throws IOException, ParseException, JSONException {
         String expectedResponse = new String(
-                Files.readAllBytes(Paths.get(System.getProperty("testDataFolderPath") +"JsonFileTest2.json")));
+                Files.readAllBytes(Paths.get(SHAFT.Properties.paths.properties() + "JsonFileTest2.json")));
         String actualResponse = new String(
-                Files.readAllBytes(Paths.get(System.getProperty("testDataFolderPath") + "JsonFileTest.json")));
+                Files.readAllBytes(Paths.get(SHAFT.Properties.paths.properties() + "JsonFileTest.json")));
 
         JSONObject expectedJsonObject = (JSONObject) (new JSONParser()).parse(expectedResponse);
         JSONObject actualJsonObject = (JSONObject) (new JSONParser()).parse(actualResponse);
