+++ src/test/java/testPackage/JsonCompareWithSpecialCharactersTests.java
@@ -1,15 +1,16 @@
 package testPackage;
 
 import com.shaft.cli.FileActions;
+import com.shaft.driver.SHAFT;
 import com.shaft.validation.Validations;
 import io.restassured.builder.ResponseBuilder;
 import io.restassured.response.Response;
 import org.testng.annotations.Test;
 
-public class Test_JsonCompareWithSpecialCharacters {
+public class JsonCompareWithSpecialCharactersTests {
     @Test
     public void specialCharacters() {
-        String referenceJsonFilePath = System.getProperty("testDataFolderPath") + "specialCharacters.json";
+        String referenceJsonFilePath = SHAFT.Properties.paths.properties() + "specialCharacters.json";
         Response response = (new ResponseBuilder()).setBody(FileActions.getInstance().readFile(referenceJsonFilePath))
                 .setStatusCode(200).build();
         Validations.verifyThat().response(response).isEqualToFileContent(referenceJsonFilePath).perform();
@@ -17,7 +18,7 @@ public class Test_JsonCompareWithSpecialCharacters {
 
     @Test
     public void specialCharacters_assertion() {
-        String referenceJsonFilePath = System.getProperty("testDataFolderPath") + "specialCharacters.json";
+        String referenceJsonFilePath = SHAFT.Properties.paths.properties() + "specialCharacters.json";
         Response response = (new ResponseBuilder()).setBody(FileActions.getInstance().readFile(referenceJsonFilePath))
                 .setStatusCode(200).build();
         Validations.assertThat().response(response).containsFileContent(referenceJsonFilePath)
