+++ src/test/resources/TestSuites/testSuite02.xml
@@ -13,7 +13,7 @@
     </test>
     <test name="Test02">
         <classes>
-            <class name="testPackage.Test_JsonActions">
+            <class name="testPackage.JsonActionsTests">
             </class>
         </classes>
     </test>
